Update controller for contact

+++ 07-nodejs-expressjs-project/controllers/contactController.js
@@ -1,8 +1,8 @@
 const asyncHandler = require('express-async-handler');
-const Contact = require('../models/contactModels');
+const Contact = require('../models/contactModel');
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
@@ -25,7 +25,12 @@ const createContact = asyncHandler(async (req, res) => {
         throw new Error('All fields are mandory');
     }
 
-    const contact = await Contact.create({ name, email, phone });
+    const contact = await Contact.create({
+        user_id: req.user.id,
+        name,
+        email,
+        phone,
+    });
 
     res.status(201).json(contact);
 });
@@ -36,6 +41,13 @@ const updateContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact Not Found');
+    }
+
+    if (!contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error(
+            "User don't have permission to update other user contacts"
+        );
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -53,6 +65,13 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact Not Found');
+    }
+
+    if (!contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error(
+            "User don't have permission to delete other user contacts"
+        );
     }
 
     await Contact.findByIdAndDelete(req.params.id);
